Move thread model into separate file
